sys,dll,inc: FSP_IOCTL_TRANSACT
